CompactFilterBehavior: Extract `RequestTracker` to a standalone file
